BUG: Suppress invalid compiler-generated methods
The GCC parser generates artificial declarations for methods that may be
implicitly defined.  Unfortunately it does so even when the method is
not supposed to be available, and waits until the code tries to call the
method to report an error.  For example, the code

  struct A { A(); const int a; };

should not have an implicit assignment operator.  In order to drop these
methods from the xml output, we test synthesizing them as if they were
called.  During the test, error diagnostics are suppressed but used to
discard the offending method.