BUG: Avoid running generated shell code
The gccxml_find_flags script used to generate a file in /tmp and then
evaluate it as shell code.  This creates a case which might allow an
attacker to insert code to be executed when the root user runs gccxml.
This fix removes use of /tmp and uses string processing instead of
executing generated code.

See issue #8083.