COMP: Tweak gcc option computation on solaris
Building on opensolaris failed because options.c did not provide the
flag_pic symbol.  It was getting preprocessed out due to an interaction
between the opt-gather.awk and optc-gen.awk scripts.  This change hacks
the latter script to make sure the symbol is provided.  Further
investigation would be required to determine the real fix, but this is
good enough for our needs.