BUG: Fix gcc option generation on opensolaris
The opensolaris native awk does not execute opt-gather.awk correctly.
This causes the incorrect generation of options.c and causes the command
line handling of gccxml_cc1plus to fail because options are sorted
incorrectly.  The solution is to prefer GNU awk if it is available.
This change also removes the work-around for flag_pic done recently
because we want the build to fail if the wrong awk is used.