BUG: Fix recursive method synthesization tests
When we synthesize an artificial method while suppressing errors to
detect if the method is valid, it may recursively synthesize other
methods.  We need to mark each method with success or failure as it is
synthesized so it can be tested when the method is encountered by the
xml dump.  See issue #8635.