BUG: Avoid ICE with dllimport-ed vtable
This works around a subtle problem in our GCC version which ICEs:

  class A {};
  struct __attribute__((dllimport)) B: virtual public A { ~B() {} };

The VAR_DECL for B's vtable is given a mark for runtime initialization,
which is not allowed for vtables.  We work around the problem by
disabling the mark completely.  It only seems to be used for code
generation, which gccxml does not need.